🐛fix (measurement): traffic struct now use uint64

+++ common/entity/traffic.go
@@ -6,7 +6,7 @@ type Traffic struct {
 	Interface   Interface `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 	InterfaceID uint
 	Date        time.Time
-	Bandwidth   uint
-	In          uint
-	Out         uint
+	Bandwidth   uint64
+	In          uint64
+	Out         uint64
 }
